Remove og:url
`og:url` should always be the URL of the page being viewed. If it is not Facebook and other content platforms will interpret this as a redirect. For example, if I wanted to share a link on Facebook about a specific speaker, my tweet would end up linking to the ViewSource homepage and not the actual speaker page.

The easiest way to rectify this is to remove the `og:url`, but ideally this would be dynamic and contain the URL of the current page.