Remove tidy, it's breaking the js
After adding a new SVG include tidy has started breaking the javascript.
I can't tell what's going on but I'd rather have working JS than tidy
source.

We were using a version of tidy without support for SVG and hacking it
by adding the SVG elements to a "new element" list. Possibly this would
not be a problem if we updated our version of tidy.